+++ app/Http/Controllers/Api/v1/Settings/ProductServiceController.php
@@ -5,10 +5,16 @@ namespace App\Http\Controllers\Api\v1\Settings;
 use App\Http\Controllers\Controller;
 use App\Models\ProductServices;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Validator;
 use RestResponse;
 class ProductServiceController extends Controller
 {
+    private $perProductServices;
+    public function __construct()
+    {
+        $this->perProductServices = config('constant.PERMISSION_PRODUCT_SERVICES_CRUD');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -17,11 +23,15 @@ class ProductServiceController extends Controller
     public function index()
     {
         try{
-            $getAllService = ProductServices::all();
-            if(empty($getAllService)){
-                return RestResponse::warning('Product service not found.');
+            if(Auth::user()->hasPermissionTo($this->perProductServices)) {
+                $getAllService = ProductServices::all();
+                if(empty($getAllService)){
+                    return RestResponse::warning('Product service not found.');
+                }
+                return RestResponse::success($getAllService,'Product service list retrieve successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            return RestResponse::success($getAllService,'Product service list retrieve successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -46,17 +56,21 @@ class ProductServiceController extends Controller
     public function store(Request $request)
     {
         try{
-            $validate = Validator::make($request->all(), [
-                'service_name' => 'required|unique:product_services,service_name,NULL,id,deleted_at,NULL'
-            ]);
-            if ($validate->fails()) {
-                return RestResponse::validationError($validate->errors());
+            if(Auth::user()->hasPermissionTo($this->perProductServices)) {
+                $validate = Validator::make($request->all(), [
+                    'service_name' => 'required|unique:product_services,service_name,NULL,id,deleted_at,NULL'
+                ]);
+                if ($validate->fails()) {
+                    return RestResponse::validationError($validate->errors());
+                }
+                $create = ProductServices::create(['service_name' => $request['service_name']]);
+                if(!$create){
+                    return RestResponse::warning('Product service create failed.');
+                }
+                return RestResponse::success([], 'Product service created successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            $create = ProductServices::create(['service_name' => $request['service_name']]);
-            if(!$create){
-                return RestResponse::warning('Product service create failed.');
-            }
-            return RestResponse::success([], 'Product service created successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -82,11 +96,15 @@ class ProductServiceController extends Controller
     public function edit($id)
     {
         try{
-            $getProductService = ProductServices::find($id);
-            if(empty($getProductService)){
-                return RestResponse::warning('Product service not found.');
+            if(Auth::user()->hasPermissionTo($this->perProductServices)) {
+                $getProductService = ProductServices::find($id);
+                if(empty($getProductService)){
+                    return RestResponse::warning('Product service not found.');
+                }
+                return RestResponse::success($getProductService,'Product service retrieve successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            return RestResponse::success($getProductService,'Product service retrieve successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -102,23 +120,27 @@ class ProductServiceController extends Controller
     public function update(Request $request, $id)
     {
         try{
-            $validate = Validator::make($request->all(), [
-                'service_name' => 'required|unique:product_services,service_name,'.$id.'NULL,id,deleted_at,NULL'
-            ]);
-            if ($validate->fails()) {
-                return RestResponse::validationError($validate->errors());
-            }
+            if(Auth::user()->hasPermissionTo($this->perProductServices)) {
+                $validate = Validator::make($request->all(), [
+                    'service_name' => 'required|unique:product_services,service_name,'.$id.'NULL,id,deleted_at,NULL'
+                ]);
+                if ($validate->fails()) {
+                    return RestResponse::validationError($validate->errors());
+                }
 
-            $findService = ProductServices::find($id);
-            if(empty($findService)){
-                return RestResponse::warning('Product service not found.');
+                $findService = ProductServices::find($id);
+                if(empty($findService)){
+                    return RestResponse::warning('Product service not found.');
+                }
+                if($findService['is_lock'] == 1){
+                    return RestResponse::warning("You can't update default product service.");
+                }
+                $findService['service_name'] = $request['service_name'];
+                $findService->save();
+                return RestResponse::success([], 'Product service updated successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            if($findService['is_lock'] == 1){
-                return RestResponse::warning("You can't update default product service.");
-            }
-            $findService['service_name'] = $request['service_name'];
-            $findService->save();
-            return RestResponse::success([], 'Product service updated successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -133,15 +155,19 @@ class ProductServiceController extends Controller
     public function destroy($id)
     {
         try{
-            $getProductService = ProductServices::find($id);
-            if (empty($getProductService)) {
-                return RestResponse::warning('Product service not found.');
+            if(Auth::user()->hasPermissionTo($this->perProductServices)) {
+                $getProductService = ProductServices::find($id);
+                if (empty($getProductService)) {
+                    return RestResponse::warning('Product service not found.');
+                }
+                if($getProductService['is_lock'] == 1){
+                    return RestResponse::warning("You can't delete default product service.");
+                }
+                $getProductService->delete();
+                return RestResponse::Success([],'Product service deleted successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            if($getProductService['is_lock'] == 1){
-                return RestResponse::warning("You can't delete default product service.");
-            }
-            $getProductService->delete();
-            return RestResponse::Success([],'Product service deleted successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
