+++ app/Http/Controllers/Api/v1/Settings/TicketStatusController.php
@@ -5,10 +5,16 @@ namespace App\Http\Controllers\Api\v1\Settings;
 use App\Http\Controllers\Controller;
 use App\Models\TicketStatus;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Validator;
 use RestResponse;
 class TicketStatusController extends Controller
 {
+    private $perTicketStatus;
+    public function __construct()
+    {
+        $this->perTicketStatus = config('constant.PERMISSION_TICKET_STATUS_CRUD');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -17,11 +23,15 @@ class TicketStatusController extends Controller
     public function index()
     {
         try{
-            $getAllTicketStatus = TicketStatus::all();
-            if(empty($getAllTicketStatus)){
-                return RestResponse::warning('Ticket Status not found.');
+            if(Auth::user()->hasPermissionTo($this->perTicketStatus)) {
+                $getAllTicketStatus = TicketStatus::all();
+                if(empty($getAllTicketStatus)){
+                    return RestResponse::warning('Ticket Status not found.');
+                }
+                return RestResponse::success($getAllTicketStatus,'Ticket Status list retrieve successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            return RestResponse::success($getAllTicketStatus,'Ticket Status list retrieve successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -46,17 +56,21 @@ class TicketStatusController extends Controller
     public function store(Request $request)
     {
         try{
-            $validate = Validator::make($request->all(), [
-                'status_name' => 'required|unique:ticket_statuses,status_name,NULL,id,deleted_at,NULL'
-            ]);
-            if ($validate->fails()) {
-                return RestResponse::validationError($validate->errors());
+            if(Auth::user()->hasPermissionTo($this->perTicketStatus)) {
+                $validate = Validator::make($request->all(), [
+                    'status_name' => 'required|unique:ticket_statuses,status_name,NULL,id,deleted_at,NULL'
+                ]);
+                if ($validate->fails()) {
+                    return RestResponse::validationError($validate->errors());
+                }
+                $createTicketStatus = TicketStatus::create(['status_name' => $request['status_name']]);
+                if(!$createTicketStatus){
+                    return RestResponse::warning('Ticket status create failed.');
+                }
+                return RestResponse::success([], 'Ticket status created successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            $createTicketStatus = TicketStatus::create(['status_name' => $request['status_name']]);
-            if(!$createTicketStatus){
-                return RestResponse::warning('Ticket status create failed.');
-            }
-            return RestResponse::success([], 'Ticket status created successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -82,11 +96,15 @@ class TicketStatusController extends Controller
     public function edit($id)
     {
         try{
-            $getTicketStatus = TicketStatus::find($id);
-            if(empty($getTicketStatus)){
-                return RestResponse::warning('Ticket status not found.');
+            if(Auth::user()->hasPermissionTo($this->perTicketStatus)) {
+                $getTicketStatus = TicketStatus::find($id);
+                if(empty($getTicketStatus)){
+                    return RestResponse::warning('Ticket status not found.');
+                }
+                return RestResponse::success($getTicketStatus,'Ticket status retrieve successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            return RestResponse::success($getTicketStatus,'Ticket status retrieve successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -102,25 +120,29 @@ class TicketStatusController extends Controller
     public function update(Request $request, $id)
     {
         try{
-            $validate = Validator::make($request->all(), [
-                'status_name' => 'required|unique:ticket_statuses,status_name,'.$id.'NULL,id,deleted_at,NULL'
-            ]);
-            if ($validate->fails()) {
-                return RestResponse::validationError($validate->errors());
-            }
+            if(Auth::user()->hasPermissionTo($this->perTicketStatus)) {
+                $validate = Validator::make($request->all(), [
+                    'status_name' => 'required|unique:ticket_statuses,status_name,'.$id.'NULL,id,deleted_at,NULL'
+                ]);
+                if ($validate->fails()) {
+                    return RestResponse::validationError($validate->errors());
+                }
 
-            $findTicketStatus = TicketStatus::find($id);
-            if(empty($findTicketStatus)){
-                return RestResponse::warning('Ticket status not found.');
-            }
+                $findTicketStatus = TicketStatus::find($id);
+                if(empty($findTicketStatus)){
+                    return RestResponse::warning('Ticket status not found.');
+                }
 
-            if($findTicketStatus['is_lock'] == 1){
-                return RestResponse::warning("You can't update default ticket status.");
-            }
+                if($findTicketStatus['is_lock'] == 1){
+                    return RestResponse::warning("You can't update default ticket status.");
+                }
 
-            $findTicketStatus['status_name'] = $request['status_name'];
-            $findTicketStatus->save();
-            return RestResponse::success([], 'Ticket status updated successfully.');
+                $findTicketStatus['status_name'] = $request['status_name'];
+                $findTicketStatus->save();
+                return RestResponse::success([], 'Ticket status updated successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
+            }
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
@@ -135,15 +157,19 @@ class TicketStatusController extends Controller
     public function destroy($id)
     {
         try{
-            $getTicketStatus = TicketStatus::find($id);
-            if (empty($getTicketStatus)) {
-                return RestResponse::warning('Ticket status not found.');
+            if(Auth::user()->hasPermissionTo($this->perTicketStatus)) {
+                $getTicketStatus = TicketStatus::find($id);
+                if (empty($getTicketStatus)) {
+                    return RestResponse::warning('Ticket status not found.');
+                }
+                if($getTicketStatus['is_lock'] == 1){
+                    return RestResponse::warning("You can't delete default ticket status.");
+                }
+                $getTicketStatus->delete();
+                return RestResponse::Success([],'Ticket status deleted successfully.');
+            }else {
+                return RestResponse::warning(config('constant.USER_DONT_HAVE_PERMISSION'));
             }
-            if($getTicketStatus['is_lock'] == 1){
-                return RestResponse::warning("You can't delete default ticket status.");
-            }
-            $getTicketStatus->delete();
-            return RestResponse::Success([],'Ticket status deleted successfully.');
         }catch (\Exception $e) {
             return RestResponse::error($e->getMessage(), $e);
         }
